requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
python-dotenv>=1.0.0

# PDF → MD
marker-pdf>=0.2.0

# MD → PDF
pypandoc>=1.12

# 개발용
black>=24.1.0
isort>=5.13.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
